moved text_detection.py to upper directory for higher coverage in book directory
