+++ booksbridge_backend/text_detection.py
